fix:oauth id issue
The issue is that we're using the Google user ID as the username in the
JWT token, but our InMemoryUserDetailsManager doesn't know about OAuth2
users. When the JWT filter tries to validate the token and load the
user, it fails because it's looking for the Google ID in our in-memory
user store.